Final/pyproject.toml
[tool.poetry]
name = "task-management"
version = "0.1.0"
description = ""
authors = ["Temirbolat Maratuly <snake2009003@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
